\name{my.plants}
\alias{my.plants}
\docType{data}
\title{
Cranberry biparental population
}
\description{
This dataset are 60 individuals from a progeny coming from the cross of 2 cranberry plants. Six SSR markers were run, 2 in the first channel (blue), 2 in the second channel (green), 2 in the third channel (yellow) and the Roxtrash375 ladder was run in the fourth channel (red). 
}
\usage{data("my.plants")}
\format{
  The format is:
 chr "my.plants"
}
\details{
The data is basically the raw FSA files coming from the ABi machine. No more details for this data.
}
\source{
This data was generated by the Cranberry Genomics Lab.
}
\references{
Covarrubias-Pazaran G, Diaz-Garcia L, Schlautman B, Salazar W, Zalapa J. Fragma: An R package for fragment analysis. http://horticulture.wisc.edu/cggl/ZalapaLab/People.html. 2015.
}
\examples{
data(my.plants)
## look at the list structure
str(my.plants) 
}
\keyword{datasets}
